buffers: fix allocation alignment of vm->buffer_main
vlib_buffer_main_t requires an alignment of CLIB_CACHE_LINE_BYTES, but
the default allocation alignment is less than this.

The C17 spec (6.3.2.3p7) states that converting to a pointer that
isn't correctly aligned for the type is undefined behaviour. In
addition, in this case it may result in non-optimal usage of the
cache.

Fix this by using clib_mem_alloc_aligned with the correct alignment
requirement for vlib_buffer_main_t.

Change-Id: Idb64c7d93c142cbf81dc7a20997c1a63e17039f9